done bo sung , chia role, update bao mat

+++ btljavafinal/spring-boot-example/src/main/java/spring/example/controller/HomeController.java
@@ -61,9 +61,13 @@ public class HomeController {
     @GetMapping("/changePassword")
     public String showChangePasswordForm(Model model, HttpSession session) {
         User userLogin = (User) session.getAttribute("userLogin");
-        if (userLogin == null) {
-            return "redirect:/login";
-        }
+
+        // xua cai may thanh session/null khong xoa vi session == null duoc nguoi dung thi khong 
+
+            if (session == null) {
+                return "redirect:/login";
+            }
+        // end phan sua 
         model.addAttribute("user", userLogin);
         return "changePassword2";
     }
@@ -94,23 +98,36 @@ public class HomeController {
     }
 
     @GetMapping("/manage-users")
-    public String getCustomers(Model model) {
-        List<User> customers = userService.getAllCustomers();
-        model.addAttribute("customers", customers);
-        return "manage-users"; // Tên của file HTML trong thư mục templates
+    public String getCustomers(Model model, HttpSession session) {
+        // phan nay khong xoa la phan chan quyen 
+        if(userService.checkAdmin(session)){
+            List<User> customers = userService.getAllCustomers();
+            model.addAttribute("customers", customers);
+            return "manage-users"; // Tên của file HTML trong thư mục templates
+        }
+        return "redirect:/home"; 
+        // chan quyen ket thuc 
     }
 
     @PostMapping("/delete/{id}")
-    public String deleteCustomer(@PathVariable Long id) {
-        userService.deleteCustomer(id);
-        return "redirect:/home/manage-users";
+    public String deleteCustomer(@PathVariable Long id,HttpSession session) {
+        // check admin khong xoa 
+        if(userService.checkAdmin(session)){
+            userService.deleteCustomer(id);
+            return "redirect:/home/manage-users";
+        }
+        return "redirect:/home"; 
     }
 
     @GetMapping("/edit/{id}")
-    public String editCustomer(@PathVariable Long id, Model model) {
-        User customer = userService.getCustomerById(id); // Lấy thông tin khách hàng theo ID
-        model.addAttribute("customer", customer);
-        return "edit-user"; // Tên file JSP cho chỉnh sửa
+    public String editCustomer(@PathVariable Long id, Model model,HttpSession session) {
+        // check admin khong xoa 
+        if(userService.checkAdmin(session)){
+            User customer = userService.getCustomerById(id); // Lấy thông tin khách hàng theo ID
+            model.addAttribute("customer", customer);
+            return "edit-user"; // Tên file JSP cho chỉnh sửa
+        }
+        return "redirect:/home"; 
     }
 
     @PostMapping("/edit/{id}")
@@ -119,7 +136,8 @@ public class HomeController {
             @RequestParam String password,
             @RequestParam String email,
             @RequestParam String phone,
-            @RequestParam String role) {
+            @RequestParam String role,
+            HttpSession session) {
         User customer = new User();
         customer.setId(id);
         customer.setUsername(username);
@@ -127,8 +145,11 @@ public class HomeController {
         customer.setEmail(email);
         customer.setRole(role);
         customer.setPhoneNumber(phone);
-
-        userService.updateCustomer(customer); // Cập nhật thông tin khách hàng
-        return "redirect:/home/manage-users";
+        // check admin khogn xoa 
+        if(userService.checkAdmin(session)){  
+            userService.updateCustomer(customer); // Cập nhật thông tin khách hàng
+            return "redirect:/home/manage-users";
+        }
+        return "redirect:/home"; 
     }
 }
